+++ components/kyma-environment-broker/internal/storage/init.go
@@ -14,11 +14,12 @@ import (
 
 const (
 	schemaName        = "public"
-	connectionRetries = 3
+	connectionRetries = 5
 )
 
 // InitializeDatabase opens database connection and initializes schema if it does not exist
 func InitializeDatabase(connectionURL string) (*dbr.Connection, error) {
+	fmt.Println(connectionURL)
 	connection, err := waitForDatabaseAccess(connectionURL, connectionRetries)
 	if err != nil {
 		return nil, err
