--- components/kyma-environment-broker/internal/storage/testutils/database.go
@@ -1,1 +1,0 @@
-package testutils
